[3.0-pre10] Use TabooLib's itemBuilder.

+++ src/main/kotlin/me/arasple/mc/trmenu/module/display/texture/Texture.kt
@@ -146,7 +146,7 @@ class Texture(
 
             if (id is Int) {
                 builder.material = XMaterial.matchXMaterial(Material::class.java.getDeclaredMethod("getMaterial").invoke(null, id) as Material ?: Material.AIR)
-                builder.damage = data.toShort()
+                builder.damage = data
             } else {
                 val name = id.toString()
                 try {
